changed MailService class to JavaMailService and added MailService interface
