+++ src/main/java/com/masonpohler/api/service/JavaMailService.java
@@ -1,0 +1,26 @@
+package com.masonpohler.api.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mail.MailException;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Component;
+
+@Component
+public class JavaMailService implements MailService {
+
+    private JavaMailSender mailSender;
+
+    @Autowired
+    JavaMailService(JavaMailSender mailSender) {
+        this.mailSender = mailSender;
+    }
+
+    public void sendMail(String to, String subject, String body) throws MailException {
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setTo(to);
+        message.setSubject(subject);
+        message.setText(body);
+        mailSender.send(message);
+    }
+}
